packages/agents/pinkman/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pyjwt==2.9.0
pydantic==2.9.2
torch==2.4.1
